PROJECT 2 Retrieval-Augmented Generation (RAG) System/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-openai
faiss-cpu
pypdf
sentence-transformers
streamlit
python-dotenv
tiktoken
